new push

+++ src/main/java/com/Property/management/Property/Management/Service/Repository/ApartmentRepository.java
@@ -3,5 +3,8 @@ package com.Property.management.Property.Management.Service.Repository;
 import com.Property.management.Property.Management.Service.Model.Apartment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApartmentRepository    extends JpaRepository<Apartment,Integer> {
+import java.util.Optional;
+
+public interface ApartmentRepository    extends JpaRepository<Apartment,Long> {
+
 }
